Add workaround for Flywheel bike power dropout.
The Flywheel bike occasionally reports spurious zero watts power values.
The underlying cause is unknown but it seems to occur once every few
minutes. This can be annoying if you are cycling at high effort and
suddenly see your power drop to zero. A simple but effective fix is to
just repeat the previous non-zero power value whenever the current power
is zero and current cadence is greater than zero. The accepted tradeoff
is that a legitimate gradual slow down to zero watts may take an extra
second to arrive at zero watts.