Fix cadence accuracy.
There are two issues here. The first and main issue is that the cadence
value reported by Zwift and other apps was 4.8576% higher than actual.
The root cause is the conversion from milliseconds to 1/1024ths was
using the wrong (inverted) scale. The second issue is flapping, i.e.
when using bot mode set to 100 rpm, we expect Zwift to report exactly
100 rpm but it flaps between 99rpm and 100rpm. The root cause is the
crank revolution timestamps were being calculated using the system clock
on every simulated pedal stroke which is subject to some amount of
jitter. The fix is to simply calculate each crank revolution timestamp
from the previous one, by adding the current cadence interval to it.

Co-authored-by: Jeremy Klein <jeremydk@gmail.com>